Add detection of newly connected nodes to the running solver.
Does not yet use the added nodes to perform work, but detects when they are
added and then queries the performance counter of the new nodes on a background
thread.

Add a simple input command interpreter in another thread to allow the user
to add a node, or quit the demonstration.